requirements.txt
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
firebase-admin==5.2.0
Flask==2.0.3
fonttools==4.29.1
gcloud==0.17.0
google-api-core==2.7.1
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
jws==0.1.3
kaleido==0.2.1
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
msgpack==1.0.3
numpy==1.22.3
oauth2client==3.0.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
plotly==5.6.0
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==3.0.7
Pyrebase3==4.0.2
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2021.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.8
scipy==1.8.0
six==1.16.0
tenacity==8.0.1
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
